set up the attacking logic

+++ src/consts.rs
@@ -10,15 +10,15 @@ pub const Q_SIZE_DIVIDER:f32 = 4.0;
 pub const Q_MAX_AGE:usize = 5000;
 
 pub const W_STR:u32 = 10; // How much mass of food a worker can carry
-pub const W_ATT:u32 = 10;
+pub const W_ATT:f32 = 10.0;
 pub const W_BASE_MASS:f32 = 20.0;
 pub const W_SIZE_DIVIDER:f32 = 4.0;
 pub const W_MAX_HP:f32 = 5.0;
 pub const W_MAX_AGE:usize = 5000;
 
 pub const S_STR:u32 = 10;
-pub const S_ATT:u32 = 10;
-pub const S_DEF:u32 = 10;
+pub const S_ATT:f32 = 10.0;
+pub const S_DEF:f32 = 10.0;
 pub const S_BASE_MASS:f32 = 20.0; // total mass is base mass + hp
 pub const S_MAX_HP:f32 = 40.0;
 pub const S_SIZE_DIVIDER:f32 = 4.0;
@@ -31,7 +31,7 @@ pub const E_SIZE_DIVIDER:f32 = 4.0;
 pub const E_MAX_AGE:usize = 5000;
 
 pub const D_STR:u32 = 10;
-pub const D_ATT:u32 = 10;
+pub const D_ATT:f32 = 10.0;
 pub const D_DEF:u32 = 10;
 pub const D_BASE_MASS:f32 = 20.0;
 pub const D_MAX_HP:f32 = 50.0;
